+++ src/EvenireDB/RawEventHeader.cs
@@ -6,16 +6,16 @@ namespace EvenireDB
     //TODO: evaluate https://github.com/MessagePack-CSharp/MessagePack-CSharp    
     internal readonly struct RawEventHeader
     {
-        public readonly ulong EventIdTimestamp;
-        public readonly ushort EventIdSequence;
+        public readonly long EventIdTimestamp;
+        public readonly int EventIdSequence;
         public readonly long DataPosition;
         public readonly byte[] EventType; // should always be size Constants.MAX_EVENT_TYPE_LENGTH
         public readonly short EventTypeLength;
         public readonly int EventDataLength;
 
         public const int SIZE =
-            sizeof(ulong) + // event id timestamp
-            sizeof(ushort) + // event id sequence
+            sizeof(long) + // event id timestamp
+            sizeof(int) + // event id sequence
             sizeof(long) + // offset in the main stream
             Constants.MAX_EVENT_TYPE_LENGTH + // type name
             sizeof(short) + // type name length
@@ -23,8 +23,8 @@ namespace EvenireDB
         ;
 
         private const int EVENT_ID_TIMESTAMP_POS = 0;
-        private const int EVENT_ID_SEQUENCE_POS = EVENT_ID_TIMESTAMP_POS + sizeof(ulong);
-        private const int OFFSET_POS = EVENT_ID_SEQUENCE_POS + sizeof(ushort);
+        private const int EVENT_ID_SEQUENCE_POS = EVENT_ID_TIMESTAMP_POS + sizeof(long);
+        private const int OFFSET_POS = EVENT_ID_SEQUENCE_POS + sizeof(int);
         private const int EVENT_TYPE_NAME_POS = OFFSET_POS + sizeof(long);
         private const int EVENT_TYPE_NAME_LENGTH_POS = EVENT_TYPE_NAME_POS + Constants.MAX_EVENT_TYPE_LENGTH;
         private const int EVENT_DATA_LENGTH_POS = EVENT_TYPE_NAME_LENGTH_POS + sizeof(short);
@@ -32,8 +32,8 @@ namespace EvenireDB
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public readonly void ToBytes(ref byte[] buffer)
         {
-            Unsafe.As<byte, ulong>(ref buffer[EVENT_ID_TIMESTAMP_POS]) = this.EventIdTimestamp;
-            Unsafe.As<byte, ushort>(ref buffer[EVENT_ID_SEQUENCE_POS]) = this.EventIdSequence;
+            Unsafe.As<byte, long>(ref buffer[EVENT_ID_TIMESTAMP_POS]) = this.EventIdTimestamp;
+            Unsafe.As<byte, int>(ref buffer[EVENT_ID_SEQUENCE_POS]) = this.EventIdSequence;
 
             // offset in the main stream            
             Unsafe.As<byte, long>(ref buffer[OFFSET_POS]) = this.DataPosition;
@@ -54,8 +54,8 @@ namespace EvenireDB
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RawEventHeader(ReadOnlySpan<byte> data)
         {
-            this.EventIdTimestamp = BitConverter.ToUInt64(data.Slice(EVENT_ID_TIMESTAMP_POS, sizeof(ulong)));
-            this.EventIdSequence = BitConverter.ToUInt16(data.Slice(EVENT_ID_SEQUENCE_POS, sizeof(ushort)));
+            this.EventIdTimestamp = BitConverter.ToInt64(data.Slice(EVENT_ID_TIMESTAMP_POS, sizeof(long)));
+            this.EventIdSequence = BitConverter.ToInt32(data.Slice(EVENT_ID_SEQUENCE_POS, sizeof(int)));
             this.DataPosition = BitConverter.ToInt32(data.Slice(OFFSET_POS, sizeof(long)));
             this.EventType = data.Slice(EVENT_TYPE_NAME_POS, Constants.MAX_EVENT_TYPE_LENGTH).ToArray();
             this.EventTypeLength = BitConverter.ToInt16(data.Slice(EVENT_TYPE_NAME_LENGTH_POS, sizeof(short)));
